+++ mycodo/config.py
@@ -14,7 +14,7 @@ sys.path.append(os.path.abspath(os.path.dirname(__file__)))
 from config_translations import TRANSLATIONS
 
 MYCODO_VERSION = '8.2.0'
-ALEMBIC_VERSION = 'c1cb0775f7ce'
+ALEMBIC_VERSION = '55aca47c2362'
 
 #  FORCE_UPGRADE_MASTER
 #  Set True to enable upgrading to the master branch of the Mycodo repository.