+++ mycodo/databases/models/__init__.py
@@ -34,6 +34,7 @@ from .alembic_version import AlembicVersion
 from .camera import Camera
 from .controller import CustomController
 from .dashboard import Dashboard
+from .dashboard import DashboardLayout
 from .display_order import DisplayOrder
 from .function import Actions
 from .function import Conditional