+++ mycodo/mycodo_flask/routes_static.py
@@ -15,8 +15,10 @@ from flask.blueprints import Blueprint
 from mycodo.config import MYCODO_VERSION
 from mycodo.config import THEMES_DARK
 from mycodo.config_translations import TRANSLATIONS
+from mycodo.databases.models import DashboardLayout
 from mycodo.databases.models import Misc
 from mycodo.mycodo_client import DaemonControl
+from mycodo.mycodo_flask.forms import forms_dashboard
 from mycodo.mycodo_flask.routes_authentication import admin_exists
 
 blueprint = Blueprint('routes_static',
@@ -39,6 +41,10 @@ blueprint.before_request(before_request_admin_exist)
 @blueprint.context_processor
 def inject_variables():
     """Variables to send with every page request"""
+    form_dashboard = forms_dashboard.DashboardConfig()  # Dashboard configuration in layout
+
+    misc = Misc.query.first()
+
     try:
         if not current_app.config['TESTING']:
             control = DaemonControl()
@@ -50,9 +56,17 @@ def inject_variables():
                      "{err}".format(err=e))
         daemon_status = '0'
 
-    misc = Misc.query.first()
+    dashboards = []
+    for each_dash in DashboardLayout.query.all():
+        dashboards.append({
+            'dashboard_id': each_dash.unique_id,
+            'name': each_dash.name
+        })
+
     return dict(dark_themes=THEMES_DARK,
                 daemon_status=daemon_status,
+                dashboards=dashboards,
+                form_dashboard=form_dashboard,
                 hide_alert_info=misc.hide_alert_info,
                 hide_alert_success=misc.hide_alert_success,
                 hide_alert_warning=misc.hide_alert_warning,